fix(tasks): send single order anchor on board drag
Task board drag-and-drop was sending both beforeTaskId and afterTaskId
for middle-column moves, but moveTask only accepts one anchor. Prefer
beforeTaskId when available and add regression tests.

<!-- linear-issue-id:JOV-2968 -->
